+++ app/_icons/keyboard-arrow-left.tsx
@@ -1,0 +1,18 @@
+export default function KeyboardArrowLeftIcon({
+  ariaLabel,
+}: {
+  ariaLabel: string;
+}) {
+  return (
+    <svg
+      aria-label={ariaLabel}
+      xmlns="http://www.w3.org/2000/svg"
+      height="24"
+      viewBox="0 -960 960 960"
+      width="24"
+      fill="currentColor"
+    >
+      <path d="M560-240 320-480l240-240 56 56-184 184 184 184-56 56Z" />
+    </svg>
+  );
+}