+++ app/_lib/key-code.ts
@@ -1,4 +1,9 @@
 export enum KeyCode {
- KeyM = 'KeyM',
- Escape = 'Escape',
+  ArrowUp = 'ArrowUp',
+  ArrowDown = 'ArrowDown',
+  ArrowLeft = 'ArrowLeft',
+  ArrowRight = 'ArrowRight',
+  KeyM = 'KeyM',
+  Escape = 'Escape',
+  Tab = 'Tab',
 }